ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.8.2
psycopg2-binary
tqdm==4.66.5 
apscheduler==3.10.4
requests
transformers
torch
summa
protobuf>=3.20,<5